Refactor into files
